requirements.txt
requests
langchain_core
langchain_ollama
langchain_openai
python-dotenv